requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
pyc3toc==1.0.0
pyFIS==1.12.2
pyserial==3.5
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
urllib3==2.0.4
